requirements.txt
beautifulsoup4==4.8.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
emoji==0.5.4
future==0.18.2
idna==2.8
lxml==4.5.0
numpy==1.18.1
pandas==1.0.1
pycparser==2.19
pymongo==3.10.1
python-dateutil==2.8.1
python-telegram-bot==12.4.2
pytz==2019.3
random-user-agent==1.0.1
redis==3.4.1
requests==2.22.0
selenium==3.141.0
sentry-sdk==0.14.1
six==1.14.0
soupsieve==1.9.5
tornado==6.0.3
urllib3==1.25.8
